Bug fix
Previous change in document computed the next ordinal before content was changed and then passed that ordinal to ordinal shift. In this case when InsertAt is called at ordinal 0, next ordinal becomes UNDEFINED_ORDINAL before the content is inserted and becomes defined only after the content has been inserted. 

UNDEFINED_ORDINAL is ignored by the consumers of ordinal shift which resulted in the caret being left at position 0 instead of moving after content insertion in a blank document.

git-svn-id: https://subversion.assembla.com/svn/sushanth/trunk@233 93233f36-2c2b-41c2-b1a5-e0ee276c51f9

+++ TextCore/TextCore/Test/FileManipulations.cs
@@ -30,9 +30,15 @@ namespace Test
             this.UIMap.OpenSampleText();
             this.UIMap.OpenSample2XML();
 
-
-            System.IO.File.Delete(UIMap.TestDataDirectory + "Current\\SampleTextSave.txt");
-            System.Diagnostics.Debug.Assert(!System.IO.File.Exists(UIMap.TestDataDirectory + "Current\\SampleTextSave.txt"));
+            if (!System.IO.Directory.Exists(UIMap.TestDataDirectory + "Current\\"))
+            {
+                System.IO.Directory.CreateDirectory(UIMap.TestDataDirectory + "Current\\");
+            }
+            if (System.IO.File.Exists(UIMap.TestDataDirectory + "Current\\SampleTextSave.txt"))
+            { 
+                System.IO.File.Delete(UIMap.TestDataDirectory + "Current\\SampleTextSave.txt");
+                System.Diagnostics.Debug.Assert(!System.IO.File.Exists(UIMap.TestDataDirectory + "Current\\SampleTextSave.txt"));
+            }
             this.UIMap.EditSampleTextAndSave();
             System.Diagnostics.Debug.Assert(System.IO.File.Exists(UIMap.TestDataDirectory + "Current\\SampleTextSave.txt"));
